Refactored to add groups blueprint + some error handling with students having no group
