+++ templates/dashboard.html
@@ -4,7 +4,7 @@
 <link rel="stylesheet" type="text/javascript" href="{{ url_for('static',filename='copytext.js') }}">
 
 <div class="card text-center shadow p-3 mb-5 bg-white rounded">
-    <a href="{{ url_for('users.create_group')}}" class="btn btn-success" >Create a New Group</a>
+    <a href="{{ url_for('groups.create_group')}}" class="btn btn-success" >Create a New Group</a>
     <div class="card mb-1">
           <div class="card-body">
             <h1 class="group-dashboard">Your Groups</h1>
