+++ templates/groups/edit_group.html
@@ -19,7 +19,7 @@
                 {% endfor %}
             {% endif %}
         {% endwith %}
-                <form method="POST" class="form" action="{{ url_for('users.edit_group', group_id=group.id) }}">
+                <form method="POST" class="form" action="{{ url_for('groups.edit_group', group_id=group.id) }}">
                 {{ form.csrf_token }}
             <div class="form-group col mb-6">
                 {{ form.name(class_="form-control", placeholder="Name", value=group.name) }}
