+++ templates/groups/group.html
@@ -7,8 +7,8 @@
 <div class="card text-center shadow p-3 mb-5 bg-white rounded">
     <div class="card mb-1">
           <img src="{{ url_for('static', filename='media/group-avatar.png') }}" class="card-img-av" alt="Default Avatar" style="float:left;width:72px;height:73px;">
-          <a href="{{ url_for('users.edit_group', group_id=group.id)}}" id="card-button-edit" class="btn btn-success">Edit Group</a>
-          <a href="{{ url_for('users.delete_group', group_id=group.id)}}" id="card-button-del" onclick="onClick()" class="btn btn-danger">Delete Group</a>
+          <a href="{{ url_for('groups.edit_group', group_id=group.id)}}" id="card-button-edit" class="btn btn-success">Edit Group</a>
+          <a href="{{ url_for('groups.delete_group', group_id=group.id)}}" id="card-button-del" onclick="onClick()" class="btn btn-danger">Delete Group</a>
           <div class="card-body">
             <h1 class="group-dashboard">{{group.name}}</h1>
             <p class="card-text"></p>
@@ -23,8 +23,8 @@
 <div class="card text-center shadow p-3 mb-5 bg-white rounded">
     <div class="card mb-1">
           <div class="card-body">
-              <a href="{{ url_for('users.create_students', group_id=group.id) }}" id="card-button-create" class="btn btn-success">Add Students</a>
-              <a href="{{ url_for('users.download_students', group_id=group.id) }}" id="card-button" class="btn btn-primary">Download Data</a>
+              <a href="{{ url_for('groups.create_students', group_id=group.id) }}" id="card-button-create" class="btn btn-success">Add Students</a>
+              <a href="{{ url_for('groups.download_students', group_id=group.id) }}" id="card-button" class="btn btn-primary">Download Data</a>
               <h1 class="group-dashboard">Students</h1>
               <br>
               <table class="table">
