+++ users/views.py
@@ -1,25 +1,19 @@
 
 # IMPORT
-import csv
-import logging
-import os
 import re
 import smtplib
 
-from email import encoders
-from email.mime.base import MIMEBase
 from email.mime.multipart import MIMEMultipart
 from email.mime.text import MIMEText
 
-from flask import flash, current_app, Blueprint, render_template, redirect, url_for, request, session
+from flask import flash, Blueprint, render_template, redirect, url_for, request, session
 from flask_login import login_user, current_user, login_required, logout_user
 from flask_navigation import Navigation
 from wtforms.fields.core import Label
 from werkzeug.security import check_password_hash, generate_password_hash
 from datetime import datetime
 from models import User, Student, Group, Teacher, Quiz, Question, StudentQuizScores
-from users.forms import CreateGroup, RegisterForm, LoginForm, QuizForm, ChangePassword, ForgottenPassword, \
-    RegisterStudent
+from users.forms import RegisterForm, LoginForm, QuizForm, ChangePassword, ForgottenPassword
 
 from random import randint
 from app import db, app, requires_roles
@@ -141,6 +135,7 @@ def logout():
 
 @users.route('/dashboard', methods=['GET'])
 @login_required
+@requires_roles('teacher')
 def dashboard():
     groups = Group.query.filter_by(teacher_id=current_user.id).all()
     return render_template('dashboard.html', groups=groups)
@@ -154,13 +149,13 @@ def account(_username):
     if current_user.role == 'student':
         student = Student.query.get(current_user.id)
         group = Group.query.get(student.group_id)
-        teacher = None
-        students = None
-        group_average_quiz_score = None
         if group is not None:
             students = len(group.students)
             teacher = Teacher.query.get(group.teacher_id)
             group_average_quiz_score = get_group_average_quiz_score(group.id)
+        else:
+            flash('You do not seem to be part of a group, please join a group to access full functionality.', 'danger')
+            return render_template('account.html', group=None, average_quiz_score=None, quizzes_completed=None)
 
         average_quiz_score, quizzes_completed = get_student_average_quiz_score(current_user.id)
         return render_template('account.html', group=group, teacher=teacher, students=students,
@@ -168,11 +163,14 @@ def account(_username):
                                group_average_quiz_score=group_average_quiz_score)
 
     if current_user.role == 'teacher':
+        group_average_quiz_score = 0
         groups = Group.query.filter_by(teacher_id=current_user.id).all()
         students = 0
         for group in groups:
             students += len(Student.query.filter_by(group_id=group.id).all())
-        return render_template('account.html', groups=groups, teacher=current_user, students=students)
+            group_average_quiz_score = get_group_average_quiz_score(group.id)
+        return render_template('account.html', groups=groups, teacher=current_user, students=students,
+                               group_average_quiz_score=group_average_quiz_score)
 
     return render_template('account.html')
 
@@ -204,10 +202,7 @@ def change_password():
             flash("Your password has been changed", "info")
             logout_user()
             current_time = datetime.now().strftime("%H:%M:%S")
-            html = f"""\
-                    <html>
-                      <head></head>
-                      <body>
+            body = f"""\
                         <p><b>Hello {user.name}. </b></p>
                         <br>
                         <br>
@@ -216,18 +211,8 @@ def change_password():
                         <p> Time of change: {current_time}
                         <br>
                         <p>If this was not done by use, please change your website password immediately.</p>
-                      </body>
-                      <br>
-                      <footer style="position:center">
-                      <b>Climate4Kids
-                      <br>
-                      <br>
-                      <i>For a better education to the next generation</i>
-                      </b>
-                      </footer>
-                    </html>
                     """
-            send_email(user, html)
+            send_email(user, body)
             return redirect(url_for('users.login', form=LoginForm()))
         # Checks if user entered correct current password
 
@@ -242,7 +227,7 @@ def forgotten_password():
         user = User.query.filter_by(username=form.username.data).first()
         # Checks if user entered correct current password
         if not user:
-            flash("That account does not exist.", "danger")
+            flash("That user does not exist.", "danger")
             return render_template('auth/forgotten_password.html', form=form)
 
         new_pass = ""
@@ -254,26 +239,13 @@ def forgotten_password():
         # Generates hash for new password and commits to the DB
         user.password = generate_password_hash(new_pass)
         db.session.commit()
-        html = f"""\
-            <html>
-              <head></head>
-              <body>
+        body = f"""\
                 <p><b>Hello {user.name}. Your new password is {new_pass}</b></p>
                 <br>
                 <br>
                 <p>We urge you to change the password as soon as possible.</p>
-              </body>
-              <br>
-              <footer style="position:center">
-              <b>Climate4Kids
-              <br>
-              <br>
-              <i>For a better education to the next generation</i>
-              </b>
-              </footer>
-            </html>
             """
-        send_email(user, html)
+        send_email(user, body)
 
         flash("Your new temporary password has been sent to your email.", "info")
         return redirect(url_for('users.login', form=LoginForm()))
@@ -294,7 +266,8 @@ def content():
     if current_user.role == 'student':
         group_id = User.query.get(current_user.id).group_id
         group = Group.query.get(group_id)
-        key_stage = group.key_stage
+        if group:
+            key_stage = group.key_stage
     return render_template('content.html', key_stage=int(key_stage))
 
 
@@ -369,25 +342,6 @@ def quiz_questions(quiz_id):
     return render_template('quizzes/quiz_question.html', form=form)
 
 
-@users.route('/groups/<string:group_id>', methods=['GET'])
-@login_required
-def group(group_id):
-    group_obj = Group.query.get(group_id)
-    student_list = Student.query.filter_by(group_id=group_id).all()
-
-    students_info = []
-    # for each student get their Name, Last login, how many quizzes they have completed and their average quiz score
-    for student in student_list:
-        average_score, num_of_quizzes = get_student_average_quiz_score(student.id)
-        student_info = [student.name, student.last_login, num_of_quizzes, int(average_score)]
-
-        students_info.append(student_info)
-
-    return render_template('groups/group.html',
-                           group=group_obj,
-                           students_info=students_info)
-
-
 # returns the average score and number of quizzes completed by the specified student
 def get_student_average_quiz_score(student_id):
     student_quiz_scores = StudentQuizScores.query.filter_by(student_id=student_id).all()
@@ -420,213 +374,33 @@ def get_group_average_quiz_score(group_id):
         return sum_of_all_scores / num_of_scores
 
 
-@users.route('/groups/create_group', methods=['GET', 'POST'])
-@login_required
-@requires_roles('teacher')
-def create_group():
-    form = CreateGroup()
-    if form.validate_on_submit():
-        group_id = randint(111111, 999999)
-        id_check = Group.query.filter_by(id=group_id).first()
-
-        # while loop to check that the ID does not already exist
-        while id_check:
-            group_id = randint(111111, 999999)
-            id_check = Group.query.filter_by(id=group_id).first()
-
-        else:
-            group = Group(id=str(group_id), name=str(form.name.data), size=form.size.data,
-                          key_stage=form.key_stage.data, teacher_id=current_user.id)
-            group.teacher_id = current_user.id
-            db.session.add(group)
-            db.session.commit()
-            return redirect(url_for('users.create_students', group_id=group_id))
-    return render_template('groups/create_group.html', form=form)
-
-
-@users.route('/groups/<string:group_id>/create_students', methods=['GET', 'POST'])
-@login_required
-@requires_roles('teacher')
-def create_students(group_id):
-    """Allows for teachers to create student accounts, then will send the details to the
-    teachers email address.
-    """
-    form = RegisterStudent()
-    teacher = Teacher.query.get(current_user.id)
-    if form.validate_on_submit():
-        # Takes the names and splits them by line
-        names = list(filter(None, (form.names.data.strip().split("\r\n"))))
-        date = datetime.now().strftime("%d/%m/%Y %H:%M:%S")  # Current time/date
-        group = Group.query.get(group_id)
-        students = int(Student.query.filter_by(group_id=group_id).count())
-        # Validates that there aren't too many students
-        if students + len(names) > group.size:
-            flash("Too many students for this group size, please consider increasing group size.", "danger")
-            return render_template('groups/create_students.html', form=form, group_id=group_id)
-
-        # Makes file with the students name/passwords
-        filename = f'{group.name}.csv'
-        with open(filename, 'w', encoding='UTF8', newline='') as f:
-            writer = csv.DictWriter(f, fieldnames=['username', 'password'])
-            writer.writeheader()
-
-            # Commits these students to the database
-            for name in names:
-                username, password = generate_account(name)
-                student = Student(role="student", name=name, username=username, password=password,
-                                  last_login=None, registered_on=date, group_id=group_id)
-                db.session.add(student)
-                db.session.commit()
-                writer.writerow({'username': username, 'password': password})
-
-        if not send_student_data(group.name, filename, teacher.email):
-            flash(f"Unable to send CSV to {teacher.email}, please contact help.Climate4kids@gmail.com", "danger")
-        else:
-            return redirect(url_for('users.group', group_id=group_id))
-
-    return render_template('groups/create_students.html', form=form, group_id=group_id)
-
-
-@users.route('/groups/<string:group_id>/delete_group', methods=['GET'])
-@login_required
-@requires_roles('teacher')
-def delete_group(group_id):
-    db.session.query(Group).filter_by(id=group_id).delete()
-    db.session.commit()
-    return redirect(url_for('users.dashboard'))
-
-
-@users.route('/groups/<string:group_id>/edit_group', methods=['GET', 'POST'])
-@login_required
-@requires_roles('teacher')
-def edit_group(group_id):
-    """Allows for teachers to edit group details.
-    """
-    form = CreateGroup()
-    group = Group.query.get(group_id)
-    if form.validate_on_submit():
-        values = {'name': form.name.data, 'size': form.size.data, 'key_stage': form.key_stage.data}
-        students = db.session.query(Student).filter_by(group_id=group_id).all()
-        if int(form.size.data) < len(students):
-            flash("You must set a size greater than or equal to the number of students in this group.", "danger")
-            return render_template('groups/edit_group.html', form=form, group_id=group_id, group=group)
-
-        db.session.query(Group).filter_by(id=group_id).update(values)
-        db.session.commit()
-        return redirect(url_for('users.group', group_id=group_id))
-    return render_template('groups/edit_group.html', form=form, group_id=group_id, group=group)
-
-
-@users.route('/groups/<string:group_id>/', methods=['GET'])
-@login_required
-@requires_roles('teacher')
-def download_students(group_id):
-    """ Method to be able to download student data, currently downloads 'name' and 'last_login',
-    could be expanded further to download quiz data when the feature is available."""
-    group = Group.query.get(group_id)
-    filename = f'{group.name}.csv'
-    with open(filename, 'w', encoding='UTF8', newline='') as f:
-        writer = csv.DictWriter(f, fieldnames=['username', 'last_login', 'quizzes_completed', 'avg_quiz_score'])
-        writer.writeheader()
-
-        for student in Student.query.filter_by(group_id=group_id):
-            writer.writerow({'username': student.username,
-                             'last_login': student.last_login,
-                             'quizzes_completed': None,
-                             'avg_quiz_score': None})
-
-    def generate():
-        """ Generation method using yield to allow for storage of the file in memory,
-        allowing the file to be removed from the system before the data is returned. """
-        with open(filename) as f:
-            yield from f
-
-        os.remove(filename)
-
-    response = current_app.response_class(generate(), mimetype='text/csv')
-    response.headers.set('Content-Disposition', 'attachment', filename=filename)
-    return response
-
-
-def send_student_data(group_name, filename, teacher_email):
-    """ Function to send the user an email containing a CSV file of the student data."""
-    message = MIMEMultipart()
-    message["Subject"] = f"{group_name} Accounts"
-    message["From"] = SMTP_EMAIL
-    message["To"] = teacher_email
-
-    html = f"""\
-    <html>
-      <head></head>
-      <body>
-        <p><b>Your student accounts created for {group_name}.</b></p>
-        <br>
-        <br>
-        <p>We urge you to change these passwords as soon as possible.</p>
-      </body>
-      <br>
-      <footer style="position:center">
-      <b>Climate4Kids
-      <br>
-      <br>
-      <i>For a better education to the next generation</i>
-      </b>
-      </footer>
-    </html>
-    """
-    message.attach(MIMEText(html, "html"))
-
-    with open(filename, "rb") as a:
-        attachment = MIMEBase("application", "octet-stream")
-        attachment.set_payload(a.read())
-    encoders.encode_base64(attachment)
-    attachment.add_header(
-        "Content-Disposition",
-        f"attachment; filename= {filename}",
-    )
-    message.attach(attachment)
-    try:
-        smtp_server = smtplib.SMTP_SSL('smtp.gmail.com', 465)
-        smtp_server.ehlo()
-        smtp_server.login(SMTP_EMAIL, SMTP_PASSWORD)
-        smtp_server.sendmail(SMTP_EMAIL, teacher_email, message.as_string())
-        smtp_server.close()
-        return True
-
-    except Exception as e:
-        print(e)  # TODO: add to logging
-        return False
-
-    finally:
-        if '.csv' in filename:
-            os.remove(filename)
-
-
-def generate_account(name):
-    """ Function to generate a username and a password of three random words and 4 random digits,
-    the digits in the password will be the same as in the username.
-    """
-    password = ""
-    nums = str(randint(1111, 9999))
-    username = str(name.strip().replace(" ", "")) + nums
-    username_check = User.query.filter_by(username=username).first()
-
-    while username_check:
-        nums = str(randint(1111, 9999))
-
-    for _ in range(3):
-        password += words[randint(0, len(words) - 1)].replace("\n", "")
-    password += nums
-
-    return username, password
-
-
-def send_email(user, html):
+def send_email(user, body):
     if user.role == 'student':
         group = Group.query.get(user.group_id)
-        email = Teacher.query.get(group.teacher_id).email
+        if group:
+            email = Teacher.query.get(group.teacher_id).email
+        else:
+            flash('You do not seem to have a teacher, please join a group to access email functionality.', 'danger')
+            return
     else:
         email = user.email
+
+    html = f"""\
+            <html>
+              <head></head>
+              <body>
+                {body}
+              </body>
+              <br>
+              <footer style="position:center">
+              <b>Climate4Kids
+              <br>
+              <br>
+              <i>For a better education to the next generation</i>
+              </b>
+              </footer>
+            </html>
+            """
 
     message = MIMEMultipart()
     message["Subject"] = f"{user.name} password reset"
@@ -643,5 +417,4 @@ def send_email(user, html):
 """  
 Useful explanation of querying with Inheritance: https://docs.sqlalchemy.org/en/14/orm/inheritance_loading.html
 """
-
